Sandpiper-backend/flask/pyproject.toml
[tool.poetry]
name = "sandpiper-flask"
version = "0.1.1"
description = "Flask-based API for Sandpiper Todo App backend"
authors = ["Asim <asymlabs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.1"
flask-restx = "^1.3.0"
flask-cors = "^6.0.0"
pydantic = "^2.10.6"
pydantic-settings = "^2.8.1"
rollbar = "^1.2.0"
rococo = "^1.0.33"
pyjwt = "^2.10.1"
pika = "^1.3.2"
psycopg2-binary = "^2.9.9"
dbutils = "^3.0.3"
pymysql = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
